Remove int casts that aren't needed

+++ Assets/PathfindingEngine.cs
@@ -26,8 +26,8 @@ namespace Assets
             Region reg = terrain.data.enclosingRegion;
             float nodeSize = 0.5f;
 
-            graph.Width = (int)(reg.upperCorner.x - reg.lowerCorner.x) * (int)(1 / nodeSize);
-            graph.Depth = (int)(reg.upperCorner.z - reg.lowerCorner.z) * (int)(1 / nodeSize);
+            graph.Width = (reg.upperCorner.x - reg.lowerCorner.x) * (int)(1 / nodeSize);
+            graph.Depth = (reg.upperCorner.z - reg.lowerCorner.z) * (int)(1 / nodeSize);
             graph.center = new Vector3(graph.Width / (2 * 1 / nodeSize), -10f, graph.Depth / (2 * 1 / nodeSize));
             graph.nodeSize = nodeSize;
